+++ src/main/java/org/ays/emergency_application/service/impl/EmergencyEvacuationApplicationServiceImpl.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.ays.auth.model.AysIdentity;
 import org.ays.common.model.AysPage;
+import org.ays.emergency_application.exception.EmergencyEvacuationApplicationNotExistException;
 import org.ays.emergency_application.model.EmergencyEvacuationApplication;
 import org.ays.emergency_application.model.filter.EmergencyEvacuationApplicationFilter;
 import org.ays.emergency_application.model.mapper.EmergencyEvacuationApplicationRequestToDomainMapper;
@@ -13,7 +14,6 @@ import org.ays.emergency_application.model.request.EmergencyEvacuationApplicatio
 import org.ays.emergency_application.port.EmergencyEvacuationApplicationReadPort;
 import org.ays.emergency_application.port.EmergencyEvacuationApplicationSavePort;
 import org.ays.emergency_application.service.EmergencyEvacuationApplicationService;
-import org.ays.emergency_application.util.exception.EmergencyEvacuationApplicationNotExistException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
